requirements.txt
# Core data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=14.0.0

# Validation & API
pydantic>=2.0.0
jsonschema>=4.20.0

# API server & clients
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
httpx>=0.25.0

# PDF & web
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0

# Config & env
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

# Optional: OpenAI for LLM report (fallback to mock if not set)
openai>=1.0.0
